Add absolute path for gpg in .gitconfig
On macOS, attempting to commit with signoff in GUI Apps, e.g.,
JetBrains IDEs result in this error:

fatal: cannot run gpg: No such file or directory
error: gpg failed to sign the data
fatal: failed to write commit object

This is because gpg is installed into /usr/local/bin/gpg (via Homebrew),
and /usr/local/bin is not in the PATH for apps launched via launchd,
like anything out of the Dock.

Add the absolute path for gpg in git config file fixes this issue.

Ref: https://youtrack.jetbrains.com/issue/IDEA-127802